introduce DataMixin, DataChangeNotifierMixin, ChildrenChangeNotifierMixin
